+++ Magic/CHANGELOG.md
@@ -11,6 +11,7 @@
  - Add `CheckChatClick` action, which can send a message to a player and then react to a click
  - Add `same_block: true` option to `Teleport` action, to allow teleporting less than one block.
    This is normally disallowed to prevent glitching through panes and iron bars
+ - Add `use_target` parameter to CheckRequirements action, to make it easier to check target requirements
 
 ## Changes